Fix delete all chat logs button.

+++ src/pages/07_settings.py
@@ -1,10 +1,18 @@
 import streamlit as st
+import const
 from modules.database import database
+from modules import common
 
 st.title("Settings")
 
 db = database.Database()
 
-if st.button("Delete all chat logs"):
-    db.delete_all_chat_logs()
-    st.success("All chat logs deleted successfully.")
+if (
+    common.check_if_exists_in_session(const.SESSION_INFO_AUTH_STATUS)
+    and st.session_state[const.SESSION_INFO_AUTH_STATUS]
+):
+    if st.button("Delete all chat logs"):
+        db.delete_all_chat_logs()
+        st.success("All chat logs deleted successfully.")
+else:
+    st.error("You are not logged in. Please go to the login page.")
